requirements.txt
Flask==2.1.3
Werkzeug==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
python-dotenv
boto3
prometheus_flask_exporter